+++ make_plugin_proto.sh
@@ -9,7 +9,9 @@ OUTDIR=compiler-plugin/src/main/java
 TMPDIR=$(mktemp -d)
 
 # Setup protoc
-PROTOC_JAR_VERSION=3.7.1
+PROTOC_JAR_VERSION=3.8.0
+COMPILER_VERSION=-v371
+
 JAR_FILENAME=protoc-jar-$PROTOC_JAR_VERSION.jar
 JAR_PATH=${HOME}/.ivy2/cache/com.github.os72/protoc-jar/jars/$JAR_FILENAME
 JAR_URL=http://central.maven.org/maven2/com/github/os72/protoc-jar/$PROTOC_JAR_VERSION/$JAR_FILENAME
@@ -21,7 +23,7 @@ fi
 
 function protoc_jar() 
 {
-  java -jar $JAR_PATH -v371 "$@"
+  java -jar $JAR_PATH "$@"
 }
 
 # SBT 1.x depends on scalapb-runtime which contains a compiled copy of
@@ -33,14 +35,18 @@ function protoc_jar()
 sed 's/scalapb\.options/scalapb.options.compiler/' \
     ./protobuf/scalapb/scalapb.proto > $TMPDIR/scalapb.proto
 
-protoc_jar --java_out="$OUTDIR" \
+protoc_jar \
+    ${COMPILER_VERSION} \
+    --java_out="$OUTDIR" \
     --proto_path=$TMPDIR \
     --proto_path=./third_party \
     $TMPDIR/scalapb.proto
 
 rm -rf $TMPDIR
 
-protoc_jar --java_out=scalapb-runtime/jvm/src/main/java --proto_path=./protobuf \
+protoc_jar \
+    -v380 \
+    --java_out=scalapb-runtime/jvm/src/main/java --proto_path=./protobuf \
     --proto_path=./third_party \
     ./protobuf/scalapb/scalapb.proto
 
@@ -61,4 +67,3 @@ sbt "++2.12.9" "scalapbc/run --scala_out=java_conversions:scalapb-runtime/jvm/sr
     "
 
 echo Done!
-