+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 # Change Log
 
-## [0.9.1] (Unreleased)
+## [0.9.1 (Unreleased)](https://github.com/scalapb/ScalaPB/tree/v0.9.0...HEAD)
 - It is now possible to specify no_box at the message-level.
 - Updated to Scala 2.12.9
 - Added `NonEmpty` marker trait to non-empty cases of sealed oneofs. A new
   method `asNonEmpty` returns an `Option[NonEmpty]`. See #633
+- Update to protobuf 3.8.0
 
-## [0.9.0](https://github.com/scalapb/ScalaPB/tree/HEAD)
+## [0.9.0](https://github.com/scalapb/ScalaPB/tree/v0.9.0)
+[Full Changelog](https://github.com/scalapb/ScalaPB/compare/v0.8.4...v0.9.0)
 - Experimental support for Scala 2.13.0. Notes:
   - ScalaPB now generates `scala.Seq` by default for repeated
     fields (previously was `scala.collection.Seq`). This ensures usage of